[DDP] Support not all outputs used in loss calculation
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57081

Changes in this diff:
1. Enable passthrough autograd function when find_unused_parameters=True.
2. With above, move `prepare_for_backward` which does unused parameter checking logic to beginning of backwards pass, only when `find_unused_parameters=True`.
3. Enhance process of unused parameter checking to account for outputs not being used in loss.

The way (3) is implemented is by triggering the autograd hook corresponding to parameters that did not participate in loss computation. Since they did not participate, the autograd hook is triggered with a gradient of zero, and if the gradient output corresponding to a model output is zero it means that it was not involved in loss computation (it could also mean that the gradient was indeed all zeros, but this is practically equivalent).
Tested by ensuring that when a model output is not used in loss, the corresponding grad is not modified. Also verified that the grads are the same in local vs DDP training case.

Note that in this diff we are not enabling the pass through autograd function for regular case find_unused_parameters=False because that has a much bigger blast radius and needs additional careful analysis especially with regard to the performance.

ghstack-source-id: 129310120

Differential Revision: [D28048628](https://our.internmc.facebook.com/intern/diff/D28048628/)